Improved motor speed and tested precision
Also removed print statements, reorganized controlAlgorithm(), increased buffer around edge of usable area, and added new test points

+++ prototype_ring_design.cydsn/i2cFunctions.h
@@ -20,7 +20,7 @@
     	int16 counter; // NOT a uint8
     	int index;
         int calibrationSteps;
-        int stepSpeed;
+        float stepSpeed;
         float lineLength;
         float nextLineLength;
         float deltaL;
